+++ client/src/cluster.rs
@@ -25,13 +25,19 @@ impl FromStr for Cluster {
             "d" | "devnet" => Ok(Cluster::Devnet),
             "l" | "localnet" => Ok(Cluster::Localnet),
             "g" | "debug" => Ok(Cluster::Debug),
-            _ if s.starts_with("http") => {
+            _ => {
                 let http_url = s;
 
                 // Taken from:
                 // https://github.com/solana-labs/solana/blob/aea8f0df1610248d29d8ca3bc0d60e9fabc99e31/web3.js/src/util/url.ts
 
                 let mut ws_url = Url::parse(http_url)?;
+                if !matches!(ws_url.scheme(), "http" | "https") {
+                    return Err(anyhow::Error::msg(
+                        "Cluster must be one of [localnet, testnet, mainnet, devnet] or be an \
+                         http or https url\n",
+                    ));
+                }
                 if let Some(port) = ws_url.port() {
                     let ws_port = port
                         .checked_add(1)
@@ -52,10 +58,6 @@ impl FromStr for Cluster {
 
                 Ok(Cluster::Custom(http_url.to_string(), ws_url.to_string()))
             }
-            _ => Err(anyhow::Error::msg(
-                "Cluster must be one of [localnet, testnet, mainnet, devnet] or be an http or \
-                 https url\n",
-            )),
         }
     }
 }
@@ -122,6 +124,12 @@ mod tests {
     }
 
     #[test]
+    fn test_reject_non_http_scheme_with_http_prefix() {
+        let bad_url = "httpx://my_custom_url.test.net";
+        assert!(Cluster::from_str(bad_url).is_err());
+    }
+
+    #[test]
     fn test_http_port() {
         let url = "http://my-url.com:7000/";
         let cluster = Cluster::from_str(url).unwrap();
@@ -173,6 +181,16 @@ mod tests {
         let cluster = Cluster::from_str(url).unwrap();
         assert_eq!(
             Cluster::Custom(url.to_string(), "ws://my-url.com/FooBar".to_string()),
+            cluster
+        );
+    }
+
+    #[test]
+    fn test_upper_case_scheme() {
+        let url = "HTTPS://my-url.com/";
+        let cluster = Cluster::from_str(url).unwrap();
+        assert_eq!(
+            Cluster::Custom(url.to_string(), "wss://my-url.com/".to_string()),
             cluster
         );
     }