+++ lang/syn/src/codegen/error.rs
@@ -21,7 +21,7 @@ pub fn generate(error: Error) -> proc_macro2::TokenStream {
             let display_msg = match &error_code.msg {
                 None => {
                     quote! {
-                        <Self as std::fmt::Debug>::fmt(self, fmt)
+                        <Self as ::std::fmt::Debug>::fmt(self, fmt)
                     }
                 }
                 Some(msg) => {
@@ -60,7 +60,7 @@ pub fn generate(error: Error) -> proc_macro2::TokenStream {
     };
 
     let ret = quote! {
-        #[derive(std::fmt::Debug, Clone, Copy)]
+        #[derive(::std::fmt::Debug, Clone, Copy)]
         #[repr(u32)]
         #error_enum
 
@@ -93,8 +93,8 @@ pub fn generate(error: Error) -> proc_macro2::TokenStream {
             }
         }
 
-        impl std::fmt::Display for #enum_name {
-            fn fmt(&self, fmt: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
+        impl ::std::fmt::Display for #enum_name {
+            fn fmt(&self, fmt: &mut ::std::fmt::Formatter<'_>) -> ::std::result::Result<(), ::std::fmt::Error> {
                 match self {
                     #(#display_variant_dispatch),*
                 }