+++ tests/declare-program/programs/declare-program/tests/generics.rs
@@ -9,10 +9,42 @@ fn structs() {
 }
 
 #[test]
+fn structs_can_clone() {
+    let strct = GenericStruct { field: [1; 2] };
+    let _ = strct.clone();
+}
+
+#[test]
+fn structs_can_debug() {
+    let strct = GenericStruct { field: [1; 2] };
+    println!("{strct:?}");
+}
+
+#[test]
 fn args() {
     let _ = external::client::args::TestCompilationGenericTypes {
         generic_struct: GenericStruct {
             field: Default::default(),
         },
     };
+}
+
+#[test]
+fn args_can_clone() {
+    let args = external::client::args::TestCompilationGenericTypes {
+        generic_struct: GenericStruct {
+            field: Default::default(),
+        },
+    };
+    let _ = args.clone();
+}
+
+#[test]
+fn args_can_debug() {
+    let args = external::client::args::TestCompilationGenericTypes {
+        generic_struct: GenericStruct {
+            field: Default::default(),
+        },
+    };
+    println!("{args:?}");
 }